# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Platform Engineering Control Plane API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams (admin sees all, user sees own teams)
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a new team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamBody"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /teams/{teamId}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team details
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamBody"
      responses:
        "200":
          description: Team updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Team deleted
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /teams/{teamId}/members:
    get:
      operationId: listTeamMembers
      tags: [teams]
      summary: List members of a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: addTeamMember
      tags: [teams]
      summary: Add a member to a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTeamMemberBody"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /teams/{teamId}/members/{userId}:
    delete:
      operationId: removeTeamMember
      tags: [teams]
      summary: Remove a member from a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Member removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        githubId:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, member]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - githubId
        - username
        - displayName
        - avatarUrl
        - email
        - role
        - createdAt
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        memberCount:
          type: integer
        namespaceCount:
          type: integer
      required:
        - id
        - name
        - slug
        - description
        - createdBy
        - createdAt
        - updatedAt
        - memberCount
        - namespaceCount
    CreateTeamBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
      required:
        - name
        - slug
    UpdateTeamBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        userId:
          type: integer
        role:
          type: string
          enum: [owner, member]
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - teamId
        - userId
        - role
        - joinedAt
        - user
    AddTeamMemberBody:
      type: object
      properties:
        userId:
          type: integer
        role:
          type: string
          enum: [owner, member]
      required:
        - userId
...